+++ tls/s2n_security_policies.h
@@ -70,8 +70,8 @@ struct s2n_security_policy {
      * https://www.rfc-editor.org/rfc/rfc8446#section-4.2.7
      */
     const struct s2n_ecc_preferences *ecc_preferences;
-    /* This field controls whether the certificate_signature_preferences also
-     * applies to local certs 
+    /* This field controls whether the certificate_signature_preferences apply 
+     * to local certs loaded on configs.
      */
     bool certificate_preferences_apply_locally;
     bool rules[S2N_SECURITY_RULES_COUNT];
@@ -196,16 +196,14 @@ int s2n_find_security_policy_from_version(const char *version, const struct s2n_
 int s2n_validate_kem_preferences(const struct s2n_kem_preferences *kem_preferences, bool pq_kem_extension_required);
 S2N_RESULT s2n_validate_certificate_signature_preferences(const struct s2n_signature_preferences *s2n_certificate_signature_preferences);
 S2N_RESULT s2n_security_policy_get_version(const struct s2n_security_policy *security_policy, const char **version);
-S2N_RESULT s2n_security_policy_validate_certificate(const struct s2n_security_policy *security_policy, const struct s2n_cert_info *info);
-
-S2N_RESULT s2n_security_policy_validate_certificate_chain(
-        const struct s2n_security_policy *security_policy,
-        struct s2n_cert_chain_and_key *cert_key_pair);
-S2N_RESULT s2n_security_policy_get_version(const struct s2n_security_policy *security_policy,
-        const char **version);
 /* Checks to see if a certificate has a signature algorithm that's in our 
  * certificate_signature_preferences list 
  */
 S2N_RESULT s2n_security_policy_validate_sig_scheme_supported(
         const struct s2n_signature_preferences *cert_sig_preferences,
         const struct s2n_cert_info *info);
+S2N_RESULT s2n_security_policy_validate_certificate(
+        const struct s2n_security_policy *security_policy, const struct s2n_cert_info *info);
+S2N_RESULT s2n_security_policy_validate_certificate_chain(
+        const struct s2n_security_policy *security_policy,
+        const struct s2n_cert_chain_and_key *cert_key_pair);
